requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt --no-strip-extras
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
django==5.0.6
    # via
    #   orbis-am-tool (pyproject.toml)
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
django-filter==24.2
    # via orbis-am-tool (pyproject.toml)
djangorestframework==3.15.1
    # via
    #   orbis-am-tool (pyproject.toml)
    #   drf-spectacular
drf-spectacular==0.27.2
    # via orbis-am-tool (pyproject.toml)
filelock==3.13.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.2.0
    # via
    #   huggingface-hub
    #   torch
gunicorn==22.0.0
    # via orbis-am-tool (pyproject.toml)
huggingface-hub==0.23.4
    # via
    #   tokenizers
    #   transformers
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jinja2==3.1.3
    # via torch
jsonschema==4.22.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown==3.6
    # via orbis-am-tool (pyproject.toml)
markupsafe==2.1.5
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
numpy==1.26.3
    # via transformers
packaging==22.0
    # via
    #   gunicorn
    #   huggingface-hub
    #   transformers
psycopg2==2.9.9
    # via orbis-am-tool (pyproject.toml)
pyyaml==6.0.1
    # via
    #   drf-spectacular
    #   huggingface-hub
    #   transformers
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.15
    # via transformers
requests==2.28.1
    # via
    #   huggingface-hub
    #   transformers
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
safetensors==0.4.3
    # via transformers
sqlparse==0.5.0
    # via django
sympy==1.12
    # via torch
tokenizers==0.19.1
    # via transformers
torch==2.3.1+cpu
    # via orbis-am-tool (pyproject.toml)
tqdm==4.64.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.41.1
    # via orbis-am-tool (pyproject.toml)
typing-extensions==4.9.0
    # via
    #   huggingface-hub
    #   torch
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.13
    # via requests
xxhash==3.4.1
    # via orbis-am-tool (pyproject.toml)
